REsueltos 6 fallos

+++ HolaSDL/EstadoPG.cpp
@@ -23,13 +23,13 @@ bool EstadoPG::onClick(){
 
 
 void EstadoPG::draw() {
-	for (int i = 0; i < objetos.size(); i++) {
+	for (size_t i = 0; i < objetos.size(); i++) {
 		objetos[i]->draw();
 	}
 }
 
 void EstadoPG::update(){
-	for (int i = 0; i < objetos.size(); i++) {
+	for (size_t i = 0; i < objetos.size(); i++) {
 		objetos[i]->update();
 	}
 }
